move physics node
